requirements.txt
asyncio==4.0.0
fastapi==0.116.1
httpx==0.28.1
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.1.1
sqlalchemy==2.0.43
uvicorn==0.35.0